requirements.txt
Django==5.0.7
psycopg2-binary==2.9.9
dj-database-url==2.2.0
whitenoise==6.7.0
gunicorn==22.0.0
python-dotenv==1.0.1